- updated exasol db version to 8.34.0 - fixed oracle connetion string using docker host from inside exasol container (tested locally on macos, now will test it in github actions) - started fixing OracleSqlDialectIT, temporarily disabled it, until the fix is finished
