+++ doc/changes/changes_3.0.6.md
@@ -18,9 +18,16 @@ Code name:
 
 #### Test Dependency Updates
 
+* Updated `com.exasol:exasol-testcontainers:7.1.3` to `7.1.6`
 * Updated `com.exasol:virtual-schema-common-jdbc:12.0.0` to `13.0.0`
 * Updated `com.exasol:virtual-schema-shared-integration-tests:3.0.0` to `3.0.1`
 * Updated `org.jacoco:org.jacoco.agent:0.8.12` to `0.8.13`
+* Added `org.junit.jupiter:junit-jupiter-api:5.13.1`
+* Added `org.junit.jupiter:junit-jupiter-params:5.13.1`
+* Removed `org.junit.jupiter:junit-jupiter:5.11.4`
+* Updated `org.mockito:mockito-junit-jupiter:5.15.2` to `5.18.0`
+* Updated `org.testcontainers:junit-jupiter:1.20.4` to `1.21.1`
+* Updated `org.testcontainers:oracle-xe:1.20.4` to `1.21.1`
 
 #### Plugin Dependency Updates
