+++ src/test/java/com/exasol/adapter/dialects/oracle/ExasolVersionCheck.java
@@ -3,7 +3,10 @@ package com.exasol.adapter.dialects.oracle;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assumptions.assumeTrue;
 
-import java.sql.*;
+import java.sql.Connection;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
 
 class ExasolVersionCheck {
