+++ src/test/java/com/exasol/adapter/dialects/oracle/IntegrationTestConstants.java
@@ -12,6 +12,8 @@ public final class IntegrationTestConstants {
     public static final int ORACLE_PORT = 1521;
     public static final String RESOURCES_FOLDER_DIALECT_NAME = "oracle";
 
+    public static final String EXASOL_VERSION = "8.34.0";
+
     private IntegrationTestConstants() {
         // intentionally left empty
     }
