+++ src/test/java/com/exasol/adapter/dialects/oracle/OracleSqlDialectIT.java
@@ -1,7 +1,7 @@
 package com.exasol.adapter.dialects.oracle;
 
 import static com.exasol.adapter.dialects.oracle.IntegrationTestConstants.*;
-import static com.exasol.adapter.dialects.oracle.OracleVirtualSchemaIntegrationTestSetup.createAdapterScript;
+import static com.exasol.adapter.dialects.oracle.OracleVirtualSchemaIntegrationTestSetup.*;
 import static com.exasol.matcher.ResultSetMatcher.matchesResultSet;
 import static com.exasol.matcher.ResultSetStructureMatcher.table;
 import static org.hamcrest.CoreMatchers.*;
@@ -10,24 +10,39 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import java.io.FileNotFoundException;
 import java.math.BigDecimal;
+import java.nio.file.Path;
 import java.sql.*;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.TimeoutException;
+import java.util.stream.Collectors;
 
 import org.hamcrest.MatcherAssert;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
+import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
+import com.exasol.bucketfs.Bucket;
+import com.exasol.bucketfs.BucketAccessException;
+import com.exasol.containers.ExasolContainer;
+import com.exasol.containers.ExasolService;
 import com.exasol.dbbuilder.dialects.exasol.*;
 import com.exasol.udfdebugging.UdfTestSetup;
 
 @Tag("integration")
 @Testcontainers
-class OracleSqlDialectIT extends AbstractOracleSqlIT {
+class OracleSqlDialectIT {
+    private static final String ORACLE_CONTAINER_NAME = IntegrationTestConstants.ORACLE_CONTAINER_NAME;
+
+    private static final String SCHEMA_ORACLE = "SCHEMA_ORACLE_" + System.currentTimeMillis();
+
+    private static final String ORACLE_JDBC_CONNECTION_NAME = "JDBC_CONNECTION";
+    private static final String ORACLE_OCI_CONNECTION_NAME = "ORACLE_CONNECTION";
 
     private static final String VIRTUAL_SCHEMA_JDBC = "VIRTUAL_SCHEMA_JDBC";
     private static final String VIRTUAL_SCHEMA_ORACLE = "VIRTUAL_SCHEMA_ORACLE";
@@ -39,47 +54,82 @@ class OracleSqlDialectIT extends AbstractOracleSqlIT {
     private static final String TABLE_ORACLE_NUMBER_HANDLING = "TABLE_ORACLE_NUMBER_HANDLING";
     private static final String TABLE_ORACLE_TIMESTAMPS = "TABLE_ORACLE_TIMESTAMPS";
 
+    @Container
+    private static final ExasolContainer<? extends ExasolContainer<?>> exasolContainer = new ExasolContainer<>(EXASOL_VERSION) //
+            .withRequiredServices(ExasolService.BUCKETFS, ExasolService.UDF).withReuse(true);
+    @Container
+    private static final OracleContainerDBA oracleContainer = new OracleContainerDBA(ORACLE_CONTAINER_NAME);
+
     private static Statement statementExasol;
 
     @BeforeAll
-    static void beforeAll() throws SQLException {
+    static void beforeAll()
+            throws BucketAccessException, TimeoutException, SQLException, FileNotFoundException, InterruptedException {
         setupOracleDbContainer();
         setupExasolContainer();
     }
 
-    private static void setupOracleDbContainer() throws SQLException {
-        final var oracleConnection = oracleContainer.createConnectionDBA("");
-        final Statement statementOracle = oracleConnection.createStatement();
-        createOracleTableAllDataTypes(statementOracle);
-        createOracleTableNumberHandling(statementOracle);
-        createOracleTableTimestamps(statementOracle);
-        createTestTablesForJoinTests(oracleContainer.createConnectionDBA(""), SCHEMA_ORACLE);
+    private static void uploadInstantClientToBucket()
+            throws BucketAccessException, TimeoutException, FileNotFoundException {
+        final Bucket bucket = exasolContainer.getDefaultBucket();
+        final String instantClientName = "instantclient-basic-linux.x64-12.1.0.2.0.zip";
+        final String instantClientPath = "src/test/resources/integration/driver/oracle";
+        bucket.uploadFile(Path.of(instantClientPath, instantClientName), "drivers/oracle/" + instantClientName);
     }
 
-    private static void setupExasolContainer() throws SQLException {
+    private static void setupExasolContainer()
+            throws BucketAccessException, TimeoutException, FileNotFoundException, SQLException, InterruptedException {
         final Connection exasolConnection = exasolContainer.createConnectionForUser(exasolContainer.getUsername(),
                 exasolContainer.getPassword());
         ExasolVersionCheck.assumeExasolVersion7(exasolConnection);
+        uploadOracleJDBCDriverToBucket(exasolContainer);
+        uploadAdapterToBucket(exasolContainer.getDefaultBucket());
+        uploadInstantClientToBucket();
         statementExasol = exasolConnection.createStatement();
 
         final UdfTestSetup udfTestSetup = new UdfTestSetup(getTestHostIpFromInsideExasol(exasolContainer),
                 exasolContainer.getDefaultBucket(), exasolConnection);
         final ExasolObjectFactory exasolFactory = new ExasolObjectFactory(exasolContainer.createConnection(""),
                 ExasolObjectConfiguration.builder().withJvmOptions(udfTestSetup.getJvmOptions()).build());
-
         final ExasolSchema schema = exasolFactory.createSchema(SCHEMA_EXASOL);
+
         final AdapterScript adapterScript = createAdapterScript(schema);
 
+        final Integer mappedPort = oracleContainer.getMappedPort(ORACLE_PORT);
         final String oracleUsername = "SYSTEM";
         final String oraclePassword = "test";
+        createOracleOCIConnection(exasolFactory, mappedPort, oracleUsername, oraclePassword);
         final ConnectionDefinition jdbcConnectionDefinition = createOracleJDBCConnection(oracleUsername, oraclePassword,
                 exasolFactory);
 
         createVirtualSchemasOnExasolDbContainer(exasolFactory, adapterScript, jdbcConnectionDefinition);
     }
 
+    private static ConnectionDefinition createOracleOCIConnection(final ExasolObjectFactory exasolFactory,
+                                                                  final Integer mappedPort, final String oracleUsername, final String oraclePassword) {
+        final String oraConnectionString = "(DESCRIPTION =" //
+                + "(ADDRESS_LIST = (ADDRESS = (PROTOCOL = TCP)" //
+                + "(HOST = " + exasolContainer.getHostIp() + " )" //
+                + "(PORT = " + oracleContainer.getOraclePort() + ")))" //
+                + "(CONNECT_DATA = (SERVER = DEDICATED)" //
+                + "(SERVICE_NAME = " + oracleContainer.getDatabaseName() + ")))";
+        return exasolFactory.createConnectionDefinition(ORACLE_OCI_CONNECTION_NAME, oraConnectionString, oracleUsername,
+                oraclePassword);
+    }
+
+    private static ConnectionDefinition createOracleJDBCConnection(final String oracleUsername,
+                                                                   final String oraclePassword, final ExasolObjectFactory exasolFactory) {
+        final String hostIp = getTestHostIpFromInsideExasol(exasolContainer);
+        final String jdbcConnectionString = "jdbc:oracle:thin:@" + hostIp + ":"
+                + oracleContainer.getOraclePort() + "/" + oracleContainer.getDatabaseName();
+
+        return exasolFactory.createConnectionDefinition(ORACLE_JDBC_CONNECTION_NAME, jdbcConnectionString,
+                oracleUsername, oraclePassword);
+
+    }
+
     private static void createVirtualSchemasOnExasolDbContainer(final ExasolObjectFactory exasolFactory,
-            final AdapterScript adapterScript, final ConnectionDefinition jdbcConnectionDefinition) {
+                                                                final AdapterScript adapterScript, final ConnectionDefinition jdbcConnectionDefinition) {
         exasolFactory.createVirtualSchemaBuilder(VIRTUAL_SCHEMA_JDBC).adapterScript(adapterScript)
                 .connectionDefinition(jdbcConnectionDefinition).properties(Map.of("SCHEMA_NAME", SCHEMA_ORACLE))
                 .build();
@@ -110,6 +160,32 @@ class OracleSqlDialectIT extends AbstractOracleSqlIT {
                         ORACLE_OCI_CONNECTION_NAME, "oracle_cast_number_to_decimal_with_precision_and_scale", "36,1"))
                 .build();
         //
+    }
+
+    private static void setupOracleDbContainer() throws SQLException {
+        final var oracleConnection = oracleContainer.createConnectionDBA("");
+        final Statement statementOracle = oracleConnection.createStatement();
+        createOracleUser(statementOracle);
+        grantAdditionalRights(statementOracle);
+        createOracleTableAllDataTypes(statementOracle);
+        createOracleTableNumberHandling(statementOracle);
+        createOracleTableTimestamps(statementOracle);
+        createTestTablesForJoinTests(oracleContainer.createConnectionDBA(""), SCHEMA_ORACLE);
+    }
+
+    private static void grantAdditionalRights(final Statement statementOracle) throws SQLException {
+        statementOracle.execute("GRANT CONNECT TO test");
+        statementOracle.execute("GRANT CREATE SESSION TO test");
+        statementOracle.execute("GRANT UNLIMITED TABLESPACE TO test");
+    }
+
+    private static void createOracleUser(final Statement statementOracle) throws SQLException {
+        final String username = SCHEMA_ORACLE;
+        final String password = SCHEMA_ORACLE;
+        statementOracle.execute("CREATE USER " + username + " IDENTIFIED BY " + password);
+        statementOracle.execute("GRANT CONNECT TO " + username);
+        statementOracle.execute("GRANT CREATE SESSION TO " + username);
+        statementOracle.execute("GRANT UNLIMITED TABLESPACE TO " + username);
     }
 
     private static void createOracleTableAllDataTypes(final Statement statementOracle) throws SQLException {
@@ -213,6 +289,31 @@ class OracleSqlDialectIT extends AbstractOracleSqlIT {
             statement.execute("INSERT INTO " + schemaName + "." + TABLE_JOIN_2 + " VALUES (2,'bbb')");
             statement.execute("INSERT INTO " + schemaName + "." + TABLE_JOIN_2 + " VALUES (3,'ccc')");
         }
+    }
+
+    private ResultSet getExpectedResultSet(final List<String> expectedColumns, final List<String> expectedRows)
+            throws SQLException {
+        final Connection connection = getExasolConnection();
+        try (final Statement statement = connection.createStatement()) {
+            final String expectedValues = expectedRows.stream().map(row -> "(" + row + ")")
+                    .collect(Collectors.joining(","));
+            final String qualifiedExpectedTableName = SCHEMA_EXASOL + ".EXPECTED";
+            statement.execute("CREATE OR REPLACE TABLE " + qualifiedExpectedTableName + "("
+                    + String.join(", ", expectedColumns) + ")");
+            statement.execute("INSERT INTO " + qualifiedExpectedTableName + " VALUES" + expectedValues);
+            return statement.executeQuery("SELECT * FROM " + qualifiedExpectedTableName);
+        }
+    }
+
+    private ResultSet getActualResultSet(final String query) throws SQLException {
+        final Connection connection = getExasolConnection();
+        try (final Statement statement = connection.createStatement()) {
+            return statement.executeQuery(query);
+        }
+    }
+
+    private Connection getExasolConnection() throws SQLException {
+        return exasolContainer.createConnection("");
     }
 
     @Test
@@ -639,7 +740,7 @@ class OracleSqlDialectIT extends AbstractOracleSqlIT {
                 "VIRTUAL_SCHEMA_ORACLE, C4, VARCHAR(50) UTF8, dddddddddddddddddddd" //
         })
         void testCharactersColumns(final String virtualSchemaName, final String columnName,
-                final String expectedColumnType, final String expectedColumnValue) {
+                                   final String expectedColumnType, final String expectedColumnValue) {
             final String qualifiedTableName = virtualSchemaName + "." + TABLE_ORACLE_ALL_DATA_TYPES;
             final String query = "SELECT " + columnName + " FROM " + qualifiedTableName;
             assertAll(() -> assertExpressionExecutionStringResult(query, expectedColumnValue),
@@ -675,7 +776,7 @@ class OracleSqlDialectIT extends AbstractOracleSqlIT {
                 "VIRTUAL_SCHEMA_ORACLE | C7 | DECIMAL(10,5) | 12345.12345" //
         }, delimiter = '|')
         void testNumberColumns(final String virtualSchemaName, final String columnName, final String expectedColumnType,
-                final String expectedValue) {
+                               final String expectedValue) {
             final String qualifiedTableName = virtualSchemaName + "." + TABLE_ORACLE_ALL_DATA_TYPES;
             final String query = "SELECT " + columnName + " FROM " + qualifiedTableName;
             assertAll(() -> assertExpressionExecutionBigDecimalResult(query, new BigDecimal(expectedValue)),
@@ -693,7 +794,7 @@ class OracleSqlDialectIT extends AbstractOracleSqlIT {
                 "VIRTUAL_SCHEMA_ORACLE | C_FLOAT126 | DOUBLE | 12345678.01234567901234567890123456789" //
         }, delimiter = '|')
         void testFloatNumbers(final String virtualSchemaName, final String columnName, final String expectedColumnType,
-                final String expectedValue) {
+                              final String expectedValue) {
             final String qualifiedTableName = virtualSchemaName + "." + TABLE_ORACLE_ALL_DATA_TYPES;
             final String query = "SELECT " + columnName + " FROM " + qualifiedTableName;
             assertAll(() -> assertExpressionExecutionFloatResult(query, Float.parseFloat(expectedValue)),
