+++ src/test/java/com/exasol/adapter/dialects/oracle/OracleVirtualSchemaIntegrationTestSetup.java
@@ -1,20 +1,24 @@
 package com.exasol.adapter.dialects.oracle;
 
-import static com.exasol.adapter.dialects.oracle.AbstractOracleSqlIT.getTestHostIpFromInsideExasol;
-import static com.exasol.adapter.dialects.oracle.IntegrationTestConstants.VIRTUAL_SCHEMAS_JAR_NAME_AND_VERSION;
-import static com.exasol.adapter.dialects.oracle.IntegrationTestConstants.VIRTUAL_SCHEMA_JAR;
+import static com.exasol.adapter.dialects.oracle.IntegrationTestConstants.*;
 import static com.exasol.dbbuilder.dialects.exasol.AdapterScript.Language.JAVA;
 
 import java.io.Closeable;
 import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.net.InetAddress;
+import java.net.NetworkInterface;
 import java.nio.file.Path;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.Statement;
+import java.util.Collections;
+import java.util.Enumeration;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.TimeoutException;
+
+import org.testcontainers.containers.GenericContainer;
 
 import com.exasol.bucketfs.Bucket;
 import com.exasol.bucketfs.BucketAccessException;
@@ -24,6 +28,7 @@ import com.exasol.dbbuilder.dialects.exasol.*;
 import com.exasol.dbbuilder.dialects.oracle.OracleObjectFactory;
 import com.exasol.drivers.JdbcDriver;
 import com.exasol.udfdebugging.UdfTestSetup;
+import com.github.dockerjava.api.model.ContainerNetwork;
 
 /**
  * This class contains the common integration test setup for all Oracle virtual schemas.
@@ -36,7 +41,7 @@ public class OracleVirtualSchemaIntegrationTestSetup implements Closeable {
 
     private final Statement oracleStatement;
     private final OracleContainerDBA oracleContainer = new OracleContainerDBA(ORACLE_CONTAINER_NAME);
-    private final ExasolContainer<? extends ExasolContainer<?>> exasolContainer = new ExasolContainer<>()
+    private final ExasolContainer<? extends ExasolContainer<?>> exasolContainer = new ExasolContainer<>(EXASOL_VERSION)
             .withRequiredServices(ExasolService.BUCKETFS, ExasolService.UDF).withReuse(true);
     private final Connection exasolConnection;
     private final Statement exasolStatement;
@@ -57,15 +62,14 @@ public class OracleVirtualSchemaIntegrationTestSetup implements Closeable {
             this.exasolStatement = this.exasolConnection.createStatement();
             this.oracleConnection = this.oracleContainer.createConnectionDBA("");
             this.oracleStatement = this.oracleConnection.createStatement();
-            final UdfTestSetup udfTestSetup = new UdfTestSetup(getTestHostIpFromInsideExasol(this.exasolContainer),
+            final UdfTestSetup udfTestSetup = new UdfTestSetup(getTestHostIpFromInsideExasol(),
                     this.exasolContainer.getDefaultBucket(), this.exasolConnection);
             this.exasolFactory = new ExasolObjectFactory(this.exasolContainer.createConnection(""),
                     ExasolObjectConfiguration.builder().withJvmOptions(udfTestSetup.getJvmOptions()).build());
             final ExasolSchema exasolSchema = this.exasolFactory.createSchema(SCHEMA_EXASOL);
             this.oracleFactory = new OracleObjectFactory(this.oracleConnection);
             this.adapterScript = createAdapterScript(exasolSchema);
-            final String connectionString = "jdbc:oracle:thin:@" + this.exasolContainer.getHostIp() + ":"
-                    + this.oracleContainer.getOraclePort() + "/" + this.oracleContainer.getDatabaseName();
+            final String connectionString = buildOracleConnectionString();
 
             this.connectionDefinition = this.exasolFactory.createConnectionDefinition("ORACLE_CONNECTION",
                     connectionString, "SYSTEM", "test");
@@ -73,6 +77,12 @@ public class OracleVirtualSchemaIntegrationTestSetup implements Closeable {
         } catch (final SQLException | BucketAccessException | TimeoutException | FileNotFoundException exception) {
             throw new IllegalStateException("Failed to created Oracle test setup.", exception);
         }
+    }
+
+    private String buildOracleConnectionString() {
+        final String hostIp = getTestHostIpFromInsideExasol();
+        return "jdbc:oracle:thin:@" + hostIp + ":"
+                + this.oracleContainer.getOraclePort() + "/" + this.oracleContainer.getDatabaseName();
     }
 
     public static void uploadOracleJDBCDriverToBucket(final ExasolContainer<? extends ExasolContainer<?>> container)
@@ -109,7 +119,7 @@ public class OracleVirtualSchemaIntegrationTestSetup implements Closeable {
     }
 
     public VirtualSchema createVirtualSchema(final String forOracleSchema,
-            final Map<String, String> additionalProperties) {
+                                             final Map<String, String> additionalProperties) {
 
         final Map<String, String> properties = new HashMap<>(Map.of("SCHEMA_NAME", forOracleSchema));
         properties.putAll(additionalProperties);
@@ -133,6 +143,57 @@ public class OracleVirtualSchemaIntegrationTestSetup implements Closeable {
             this.oracleContainer.stop();
         } catch (final SQLException exception) {
             throw new IllegalStateException("Failed to stop test setup.", exception);
+        }
+    }
+
+    private String getTestHostIpFromInsideExasol() {
+        return getTestHostIpFromInsideExasol(exasolContainer);
+    }
+
+    /**
+     * Determines the host IP address that should be used inside the Exasol container
+     * to refer to the test host (typically the machine running the integration tests).
+     * <p>
+     * This method supports two environments:
+     * <ul>
+     *     <li><b>CI environments (e.g., GitHub Actions):</b> It retrieves the Docker gateway IP address
+     *         from the Exasol container's network settings.</li>
+     *     <li><b>Local macOS environments:</b> If the environment variable {@code LOCAL_MACOS_ENV=true}
+     *         is set, it attempts to find the first available non-loopback IPv4 address of the local machine.</li>
+     * </ul>
+     *
+     * @param container the running Exasol container used to access network configuration
+     * @return the IP address as seen from inside the Exasol container, or {@code null} if it cannot be determined
+     */
+    public static String getTestHostIpFromInsideExasol(GenericContainer<?> container) {
+        String localMacosEnv = System.getenv("LOCAL_MACOS_ENV");
+
+        if (localMacosEnv == null || !localMacosEnv.equalsIgnoreCase("true")) {
+            // This works inside GitHub Actions container environment
+            final Map<String, ContainerNetwork> networks = container.getContainerInfo().getNetworkSettings().getNetworks();
+            if (networks.isEmpty()) {
+                return null;
+            }
+            return networks.values().iterator().next().getGateway();
+        } else {
+            // Fallback for local machine: find a non-loopback IPv4 address
+            try {
+                Enumeration<NetworkInterface> nets = NetworkInterface.getNetworkInterfaces();
+                for (NetworkInterface netint : Collections.list(nets)) {
+                    if (netint.isUp() && !netint.isLoopback()) {
+                        Enumeration<InetAddress> inetAddresses = netint.getInetAddresses();
+                        for (InetAddress inetAddress : Collections.list(inetAddresses)) {
+                            if (!inetAddress.isLoopbackAddress() && inetAddress instanceof java.net.Inet4Address) {
+                                return inetAddress.getHostAddress();
+                            }
+                        }
+                    }
+                }
+            } catch (Exception e) {
+                throw new IllegalStateException("Failed to determine local host IP address in macOS environment", e);
+            }
+            // If everything fails, return null or throw
+            return null;
         }
     }
 }
