+++ src/test/java/com/exasol/adapter/dialects/oracle/extension/ExtensionIT.java
@@ -8,12 +8,16 @@ import static org.hamcrest.Matchers.containsString;
 import java.io.FileNotFoundException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.sql.*;
+import java.sql.Connection;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.Collection;
 import java.util.List;
 import java.util.concurrent.TimeoutException;
 
-import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
 
 import com.exasol.adapter.RequestDispatcher;
 import com.exasol.adapter.dialects.oracle.IntegrationTestConstants;
